requirements.txt
JSON-log-formatter==0.5.2
paho-mqtt==1.6.1
